[board] Update the OpenAPI to the latest version
